+++ packages/scans/src/models/Alert.ts
@@ -1,0 +1,7 @@
+import { Instance } from "./Instance";
+
+export interface Alert {
+  name: string;
+  pluginid: string;
+  instances: Instance[];
+}
