+++ packages/scans/src/models/FilteredReport.ts
@@ -1,0 +1,6 @@
+import { Report } from "./Report";
+import { FilteredSite } from "./FilteredSite";
+
+export interface FilteredReport extends Report {
+  site: FilteredSite[];
+}
