+++ packages/scans/src/models/Instance.ts
@@ -1,0 +1,3 @@
+export interface Instance {
+  uri: string;
+}
