+++ packages/scans/src/models/Report.ts
@@ -1,0 +1,6 @@
+import { Site } from "./Site";
+
+export interface Report {
+  site: Site[];
+  updated?: unknown;
+}
