+++ packages/scans/src/models/Site.ts
@@ -1,0 +1,6 @@
+import { Alert } from "./Alert";
+
+export interface Site {
+  "@name": string;
+  alerts?: Alert[];
+}
